+++ js/mapviewgl.js
@@ -7,6 +7,7 @@ var mouse = new THREE.Vector2(), INTERSECTED;
 var mixers = [];
 var misc_followers = [];
 var MeshLoader = new THREE.GLTFLoader();
+var textLabels = [];
 var WIDTH = window.innerWidth , HEIGHT = window.innerHeight;
 window.currentActiveShips = [];
 window.drawnActiveShips = []
@@ -66,12 +67,12 @@ function init() {
 				});
 
         renderer.setSize( window.innerWidth, window.innerHeight );
-        document.body.appendChild( renderer.domElement );
-				container = document.createElement( 'div' );
+        document.getElementById("container").appendChild( renderer.domElement );
+				container = document.getElementById("container")
 				document.body.appendChild( container );
 
 
-        camera = new THREE.PerspectiveCamera(50, window.innerWidth / window.innerHeight, 1, 1e7);
+        camera = new THREE.PerspectiveCamera(55, window.innerWidth / window.innerHeight, 1, 1e7);
         controls = new THREE.OrbitControls( camera, renderer.domElement );
 				controls.enableDamping = true;
 				controls.dampingFactor = 0.25;
@@ -81,7 +82,7 @@ function init() {
 
 
 		// Add some Ambient lighting   (Removed for now as it is too strong and shadows are nice)
-		
+
 		// var lightsource = new THREE.AmbientLight( 0x404040, 150000 );
 		// Add the light to the scene
 		// scene.add( lightsource );
@@ -130,13 +131,21 @@ function init() {
 		    p_mesh.name = escapeHTML(planet.name);
 				p_mesh.add( hitbox );
 				scene.add( p_mesh );
-		    l_text = new Text2D(escapeHTML(planet.name), { align: textAlign.right,  font: '12px Arial', fillStyle: '#FFF' , antialias: false });
-		    l_text.material.alphaTest = 0.0;
-		    l_text.position.set(planet.x,planet.y,planet.z);
-		    l_text.scale.set(0.25,0.25,0.25);
-				l_text.name = escapeHTML(planet.name + "_label");
-		    scene.add(l_text);
-
+				if ( preferences.get("htmlLabels") == 'true' ) {
+					l_text = drawLabel();
+					l_text.setHTML( escapeHTML(planet.name) );
+					l_text.setOffset({x:0, y:-10})
+					l_text.setParent( p_mesh );
+					textLabels.push( l_text );
+					container.appendChild( l_text.element )
+				} else {
+			    l_text = new Text2D(escapeHTML(planet.name), { align: textAlign.right,  font: '12px Arial', fillStyle: '#FFF' , antialias: false });
+			    l_text.material.alphaTest = 0.0;
+			    l_text.position.set(planet.x,planet.y,planet.z);
+			    l_text.scale.set(0.25,0.25,0.25);
+					l_text.name = escapeHTML(planet.name + "_label");
+			    scene.add(l_text);
+				}
 		  }
 
 		  // Base Generation
@@ -151,15 +160,21 @@ function init() {
 		    s_mesh.position.z=base.z;
 				s_mesh.name = escapeHTML(base.name);
 		    scene.add( s_mesh );
-		 		l_text = new Text2D(escapeHTML(base.name), { align: textAlign.left,  font: '12px Arial', fillStyle: '#ABABAB' , antialias: false });
-		    l_text.material.alphaTest = 0.0;
-		    l_text.position.set(base.x,base.y+3,base.z);
-				l_text.scale.set(0.20,0.20,0.20);
-				l_text.name = escapeHTML(base.name + "_label");
-		    scene.add(l_text);
+				if ( preferences.get("htmlLabels") == 'true' ) {
+					l_text = drawLabel();
+					l_text.setHTML( escapeHTML(base.name) );
+					l_text.setParent( s_mesh );
+					textLabels.push( l_text );
+					container.appendChild( l_text.element )
+				} else {
+					l_text = new Text2D(escapeHTML(base.name), { align: textAlign.left,  font: '12px Arial', fillStyle: '#ABABAB' , antialias: false });
+					l_text.material.alphaTest = 0.0;
+					l_text.position.set(base.x,base.y+3,base.z);
+					l_text.scale.set(0.20,0.20,0.20);
+					l_text.name = escapeHTML(base.name + "_label");
+					scene.add(l_text);
+			}
 		  }
-
-
 
 		}
     // Set view and camera to point to initial location
@@ -210,7 +225,7 @@ function onCanvasClick( event ) {
 
 function animate() {
 				var delta = clock.getDelta();
-	      requestAnimationFrame( animate );
+				requestAnimationFrame( animate );
 				update_animations();
 	      scene.updateMatrixWorld()
 				controls.update(delta);
@@ -230,16 +245,20 @@ function update_animations() {
 }
 
 
-function render () {
+function render() {
 		//requestAnimationFrame( render );
 
-    var objectlist = Object.keys(listobjects("stations"));
-    objectlist.forEach (function(station) { var obj = scene.getObjectByName(escapeHTML(station + "_label")); obj.lookAt(camera.position)  }) ;
-    objectlist = Object.keys(listobjects("planets"));
-    objectlist.forEach (function(planet) { var obj = scene.getObjectByName(escapeHTML(planet + "_label")); obj.lookAt(camera.position)  }) ;
+    if (preferences.get('htmlLabels') == "false") {
+			var objectlist = Object.keys(listobjects("stations"));
+			objectlist.forEach (function(station) { var obj = scene.getObjectByName(escapeHTML(station + "_label")); obj.lookAt(camera.position)  }) ;
+    	objectlist = Object.keys(listobjects("planets"));
+    	objectlist.forEach (function(planet) { var obj = scene.getObjectByName(escapeHTML(planet + "_label")); obj.lookAt(camera.position)  }) ;
+		}
     objectlist = Object.keys(listobjects("borders"));
     objectlist.forEach (function(border) { var obj = scene.getObjectByName(border + "_label"); if (obj != undefined) { obj.lookAt(camera.position)}  }) ;
-
+		for(var i=0; i<textLabels.length; i++) {
+          textLabels[i].updatePosition();
+        }
     renderer.render( scene, camera );
  }
 
@@ -806,3 +825,112 @@ function findPointBorder( point=new THREE.Vector3(0,0,0) ) {
 		 });
 		 	return insideborder;
 }
+
+
+var drawLabel = function() {
+	var div = document.createElement('div');
+	var visibleText = document.createElement('span');
+	var hoverText = document.createElement('span');
+	var offset = {x:10,y:15};
+	hoverText.className = "label-hover-text";
+	div.className = 'label-text';
+	div.style.position = 'fixed';
+	div.style.width = 100;
+	div.style.height = 100;
+	div.innerHTML = "";
+	div.style.top = -1000;
+	div.style.left = -1000;
+
+	var _this = this;
+
+	div.appendChild( visibleText );
+	div.appendChild( hoverText )
+
+	return {
+		element: div,
+		hover: hoverText,
+		visible: visibleText,
+		parent: false,
+		position: new THREE.Vector3(0,0,0),
+		setHTML: function(html) {
+			this.visible.innerHTML = html;
+		},
+		setHover: function(html) {
+			this.hover.innerHTML = html;
+		},
+		setParent: function(threejsobj) {
+			this.parent = threejsobj;
+		},
+		setOffset: function(given_offset){
+			offset = given_offset;
+		},
+		updatePosition: function() {
+			if(parent) {
+				this.position.copy(this.parent.position);
+			}
+
+			var coords2d = this.get2DCoords(this.position, camera);
+			//var coords2d = _this.toScreenXY(this.position,camera,container)
+			//var coords2d = this.getScreenPosition(this.parent, camera);
+			this.element.style.left = coords2d.x + 'px';
+			this.element.style.top = coords2d.y + 'px';
+
+			this.updateVisibility(); // Don't display outside of the radar box
+		},
+		updateVisibility: function() {
+			var containerloc = container.getBoundingClientRect();
+			var radarLoc = this.element.getBoundingClientRect();
+			var top = radarLoc.top;
+			var bottom = radarLoc.bottom;
+			var left = radarLoc.left;
+			var right = radarLoc.right;
+
+
+			var distance = camera.position.distanceTo(this.parent.position)
+			if((top < containerloc.top || bottom > containerloc.bottom || left < containerloc.left || right > containerloc.right) ) {
+				this.element.style.zIndex = "-100"
+			} else {
+				this.element.style.zIndex = "1"
+				var size = Math.min(1.0, Math.max(0.5 ,(600/distance)));
+				this.element.style.fontSize = `${size}em`;
+
+				}
+		}, /* End updateVisibility */
+		get2DCoords: function(position, fcamera) {
+			//  var vector = position.project( fcamera );
+			//  vector.x = (vector.x + 1)/2 * container.clientWidth;
+			//  vector.y = -(vector.y - 1)/2 * container.clientHeight;
+			var vector = new THREE.Vector3();
+			vector = vector.setFromMatrixPosition( this.parent.matrixWorld );
+			vector.project( fcamera );
+			var widthHalf = WIDTH / 2;
+			var heightHalf = HEIGHT / 2;
+			var containerloc = container.getBoundingClientRect();
+			vector.x = ( (vector.x * widthHalf) + widthHalf )  + containerloc.left - offset.x ;
+			vector.y =  ( - (vector.y * heightHalf) + heightHalf ) + containerloc.top - offset.y;
+
+			return vector;
+		}, /* End get2DCoord */
+		getScreenPosition: function(obj, camera)    {
+			var vector = new THREE.Vector3();
+
+			var widthHalf = 0.5*renderer.context.canvas.width;
+			var heightHalf = 0.5*renderer.context.canvas.height;
+
+			obj.updateMatrixWorld();
+			vector.setFromMatrixPosition(obj.matrixWorld);
+			vector.project(camera);
+
+			vector.x = ( vector.x * widthHalf ) + widthHalf;
+			vector.y = - ( vector.y * heightHalf ) + heightHalf;
+
+			return {
+					x: vector.x,
+					y: vector.y
+			};
+
+		},
+	};
+
+
+} /* End drawLabel */